fastapi_app/app/instrumentation/tracing.py
import logging

from app.core.config import settings

logger = logging.getLogger(__name__)


def install_tracing(app):
    try:
        from opentelemetry import trace
        from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
        from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
        from opentelemetry.sdk.resources import Resource
        from opentelemetry.sdk.trace import TracerProvider
        from opentelemetry.sdk.trace.export import BatchSpanProcessor
    except Exception as exc:
        logger.warning("Tracing disabled; OpenTelemetry dependencies unavailable: %s", exc)
        return

    resource = Resource.create({"service.name": settings.otel_service_name, "deployment.environment": settings.environment})
    provider = TracerProvider(resource=resource)
    span_exporter = OTLPSpanExporter(endpoint=settings.otel_exporter_otlp_endpoint, insecure=True)
    provider.add_span_processor(BatchSpanProcessor(span_exporter))
    trace.set_tracer_provider(provider)
    FastAPIInstrumentor.instrument_app(app)
