fastapi_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
sqlalchemy==2.0.31
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
redis==5.0.7
celery==5.4.0
clickhouse-connect==0.7.16
httpx==0.27.0
prometheus-client==0.20.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
opentelemetry-instrumentation-redis==0.47b0
